requirements.txt
langchain
openai
pandas
openpyxl
dataclasses
langchain-community
python-dotenv
fastapi
uvicorn
bcrypt
